consolidate experiments
