+++ src/main/java/com/marginallyclever/robotoverlord/experiments/ClassExplorer.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.systems.robot.robotarm.robotpanel;
+package com.marginallyclever.robotoverlord.experiments;
 
 import com.marginallyclever.communications.application.TextInterfaceToListeners;
 import com.marginallyclever.communications.application.TextInterfaceToSessionLayer;
