+++ src/main/java/com/marginallyclever/robotoverlord/experiments/TapeDeckPanel.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.systems.robot.robotarm.robotpanel;
+package com.marginallyclever.robotoverlord.experiments;
 
 import com.marginallyclever.convenience.log.Log;
 import org.apache.batik.ext.swing.GridBagConstants;
